+++ libs/rucken/todo-nestjs/src/services/index.ts
@@ -4,4 +4,10 @@ import { TasksService } from './tasks.service';
 import { ProjectsUsersService } from './projects-users.service';
 import { ProjectsStatusesService } from './projects-statuses.service';
 
-export const TODO_SERVICES = [ProjectsService, StatusesService, TasksService, ProjectsUsersService, ProjectsStatusesService];
+export const TODO_SERVICES = [
+  ProjectsService,
+  StatusesService,
+  TasksService,
+  ProjectsUsersService,
+  ProjectsStatusesService
+];